full trak atom!
