+++ packages/media-parser/src/create/iso-base-media/trak/mdia/minf/stbl/stsd/create-avcc.ts
@@ -1,5 +1,5 @@
-import {combineUint8Arrays} from '../../../../../../boxes/webm/make-header';
-import {addSize, stringsToUint8Array} from '../../../../primitives';
+import {combineUint8Arrays} from '../../../../../../../boxes/webm/make-header';
+import {addSize, stringsToUint8Array} from '../../../../../primitives';
 
 export const createAvccBox = (privateData: Uint8Array) => {
 	return addSize(
