+++ packages/media-parser/src/test/btrt.test.ts
@@ -1,5 +1,5 @@
 import {expect, test} from 'bun:test';
-import {createBtrt} from '../create/iso-base-media/mdia/minf/stbl/stsd/create-btrt';
+import {createBtrt} from '../create/iso-base-media/trak/mdia/minf/stbl/stsd/create-btrt';
 
 const reference = new Uint8Array([
 	0x00, 0x00, 0x00, 0x14, 0x62, 0x74, 0x72, 0x74, 0x00, 0x00, 0x00, 0x00, 0x00,
