+++ packages/media-parser/src/test/stsc.test.ts
@@ -54,7 +54,7 @@ const reference = new Uint8Array([
 
 import {expect, test} from 'bun:test';
 import {exampleAudioSamplePositions} from '../create/iso-base-media/example-stts';
-import {createStsc} from '../create/iso-base-media/mdia/minf/stbl/create-stsc';
+import {createStsc} from '../create/iso-base-media/trak/mdia/minf/stbl/create-stsc';
 
 test('stsc', () => {
 	expect(reference.byteLength).toBe(1252);
