+++ packages/media-parser/src/test/stts.test.ts
@@ -1,6 +1,6 @@
 import {expect, test} from 'bun:test';
 import {exampleAudioSamplePositions} from '../create/iso-base-media/example-stts';
-import {createSttsAtom} from '../create/iso-base-media/mdia/minf/stbl/create-stts';
+import {createSttsAtom} from '../create/iso-base-media/trak/mdia/minf/stbl/create-stts';
 
 const reference = new Uint8Array([
 	0x00, 0x00, 0x00, 0x18, 0x73, 0x74, 0x74, 0x73, 0x00, 0x00, 0x00, 0x00, 0x00,
